Create Help For Create
Hello, dear friends and artists, I have an idea and an opinion, to make a profitable program, without special code and without the need to have a lot of time for programming! It is possible to write the program in half a day! In general, I am looking for someone who can design and build the program for the Apple system, of course the idea is mine! But how can I trust?